+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - `@bedrock/quasar@10`.
 - Unpin `quasar`.
 - Use quasar `platform.is.mobile` instead of userAgent regex.
+- Update lint tooling and fix lint issues.
 
 ## 6.0.2 - 2024-03-13